+++ src/heap.cpp
@@ -6,15 +6,15 @@
 #include <functional>
 #include <iostream>
 
-static size_t get_left(size_t pos) {
+static int get_left(int pos) {
   return 2*pos + 1;
 }
 
-static size_t get_right(size_t pos) {
+static int get_right(int pos) {
   return 2*pos + 2;
 }
 
-static size_t get_parent(size_t pos) {
+static int get_parent(int pos) {
   return (pos - 1)/2;
 }
 
@@ -61,10 +61,10 @@ void heap::insert(node * n) {
 }
 
 void heap::adjust_entry(node * n) {
-  size_t pos = n->get_heap_pos();
+  int pos = n->get_heap_pos();
   assert(_data[pos] == n);
-  assert(pos < _size);
-  size_t parent = get_parent(pos);
+  assert(pos < static_cast<int>(_size));
+  int parent = get_parent(pos);
   while (pos > 0 && get_value(parent) > get_value(pos)) {
     swap(parent, pos);
     pos = parent;
@@ -74,11 +74,11 @@ void heap::adjust_entry(node * n) {
 
 void heap::print() const {
   std::cout << std::endl << "HEAP:" << std::endl;;
-  size_t level = 0;
-  size_t i0 = 0, i1 = 1ul << level;
+  int level = 0;
+  int i0 = 0, i1 = 1ul << level;
   while (i0 < _size) {
     std::cout << level << ":";
-    for (size_t i = i0; i < std::min(_size, i1); ++i) {
+    for (int i = i0; i < std::min(static_cast<int>(_size), i1); ++i) {
       std::cout << " " << get_value(i);
     }
     std::cout << std::endl;
@@ -96,10 +96,10 @@ void heap::grow() {
   _data = tmp;
 }
 
-void heap::heapify(size_t pos) {
-  std::function<void(size_t)> const rec = [&] (size_t pos) {
-    size_t l = get_left(pos), r = get_right(pos);
-    size_t smallest;
+void heap::heapify(int pos) {
+  std::function<void(int)> const rec = [&] (int pos) {
+    int l = get_left(pos), r = get_right(pos);
+    int smallest;
     if (l < _size && get_value(l) < get_value(pos)) {
       smallest = l;
     } else {
@@ -116,13 +116,13 @@ void heap::heapify(size_t pos) {
   rec(pos);
 }
 
-void heap::swap(size_t pos1, size_t pos2) {
+void heap::swap(int pos1, int pos2) {
 	std::swap(_data[pos1], _data[pos2]);
 	_data[pos1]->set_heap_pos(pos1);
 	_data[pos2]->set_heap_pos(pos2);
 }
 
-double heap::get_value(size_t pos) const {
+double heap::get_value(int pos) const {
 	return _data[pos]->get_value();
 }
