+++ src/node.cpp
@@ -1,6 +1,6 @@
 #include "node.hpp"
 
-node node::make_boundary_node(size_t i, size_t j) {
+node node::make_boundary_node(int i, int j) {
   node n;
   n._value = 0;
   n._state = state::valid;
@@ -17,27 +17,27 @@ void node::set_value(double value) {
   _value = value;
 }
 
-size_t node::get_i() const {
+int node::get_i() const {
   return _i;
 }
 
-void node::set_i(size_t i) {
+void node::set_i(int i) {
   _i = i;
 }
 
-size_t node::get_j() const {
+int node::get_j() const {
   return _j;
 }
 
-void node::set_j(size_t j) {
+void node::set_j(int j) {
   _j = j;
 }
 
-size_t node::get_heap_pos() const {
+int node::get_heap_pos() const {
   return _heap_pos;
 }
 
-void node::set_heap_pos(size_t pos) {
+void node::set_heap_pos(int pos) {
   _heap_pos = pos;
 }
