+++ src/olim8_rhr.cpp
@@ -6,7 +6,7 @@
 
 #include "olim8_util.hpp"
 
-void olim8_rhr::update_node_value_impl(size_t i, size_t j, double & T) {
+void olim8_rhr::update_node_value_impl(int i, int j, double & T) {
   node * nb[8] = {0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0};
   node * x0 = 0x0;
   node * x1 = 0x0;
